--- src/main/java/lotto/committee/drawstation/DrawMachine.java
@@ -1,9 +1,0 @@
-package lotto.committee.drawstation;
-
-import java.util.List;
-
-public class DrawMachine {
-
-    private final List
-
-}
