+++ src/main/java/lotto/drawsystem/DrawMachine.java
@@ -1,0 +1,14 @@
+package lotto.drawsystem;
+
+import java.util.TreeSet;
+
+public class DrawMachine {
+
+    private Integer drawnMainNumber;
+    private Integer drawnBonusNumber;
+    private final TreeSet<Integer> mainNumbers;
+    private final Integer bonusNumber;
+
+
+
+}
